reformat all perl files according to perltidyrc

+++ Bugzilla/Auth/Persist/Cookie.pm
@@ -26,14 +26,15 @@ sub new {
 }
 
 sub persist_login {
-    my ($self, $user) = @_;
-    my $dbh = Bugzilla->dbh;
-    my $cgi = Bugzilla->cgi;
+    my ( $self, $user ) = @_;
+    my $dbh          = Bugzilla->dbh;
+    my $cgi          = Bugzilla->cgi;
     my $input_params = Bugzilla->input_params;
 
     my $ip_addr;
-    if ($input_params->{'Bugzilla_restrictlogin'}) {
+    if ( $input_params->{'Bugzilla_restrictlogin'} ) {
         $ip_addr = remote_ip();
+
         # The IP address is valid, at least for comparing with itself in a
         # subsequent login
         trick_taint($ip_addr);
@@ -41,18 +42,18 @@ sub persist_login {
 
     $dbh->bz_start_transaction();
 
-    my $login_cookie = 
-        Bugzilla::Token::GenerateUniqueToken('logincookies', 'cookie');
+    my $login_cookie = Bugzilla::Token::GenerateUniqueToken( 'logincookies', 'cookie' );
 
-    $dbh->do("INSERT INTO logincookies (cookie, userid, ipaddr, lastused)
+    $dbh->do(
+        "INSERT INTO logincookies (cookie, userid, ipaddr, lastused)
               VALUES (?, ?, ?, NOW())",
-              undef, $login_cookie, $user->id, $ip_addr);
+        undef, $login_cookie, $user->id, $ip_addr
+    );
 
     # Issuing a new cookie is a good time to clean up the old
     # cookies.
-    $dbh->do("DELETE FROM logincookies WHERE lastused < "
-             . $dbh->sql_date_math('LOCALTIMESTAMP(0)', '-',
-                                   MAX_LOGINCOOKIE_AGE, 'DAY'));
+    $dbh->do( "DELETE FROM logincookies WHERE lastused < "
+            . $dbh->sql_date_math( 'LOCALTIMESTAMP(0)', '-', MAX_LOGINCOOKIE_AGE, 'DAY' ) );
 
     $dbh->bz_commit_transaction();
 
@@ -61,45 +62,50 @@ sub persist_login {
     return $login_cookie if i_am_webservice();
 
     # Prevent JavaScript from accessing login cookies.
-    my %cookieargs = ('-httponly' => 1);
+    my %cookieargs = ( '-httponly' => 1 );
 
     # Remember cookie only if admin has told so
     # or admin didn't forbid it and user told to remember.
-    if ( Bugzilla->params->{'rememberlogin'} eq 'on' ||
-         (Bugzilla->params->{'rememberlogin'} ne 'off' &&
-          $input_params->{'Bugzilla_remember'} &&
-          $input_params->{'Bugzilla_remember'} eq 'on') ) 
+    if (Bugzilla->params->{'rememberlogin'} eq 'on'
+        || (   Bugzilla->params->{'rememberlogin'} ne 'off'
+            && $input_params->{'Bugzilla_remember'}
+            && $input_params->{'Bugzilla_remember'} eq 'on' )
+        )
     {
         # Not a session cookie, so set an infinite expiry
         $cookieargs{'-expires'} = 'Fri, 01-Jan-2038 00:00:00 GMT';
     }
-    if (Bugzilla->params->{'ssl_redirect'}) {
-        # Make these cookies only be sent to us by the browser during 
+    if ( Bugzilla->params->{'ssl_redirect'} ) {
+
+        # Make these cookies only be sent to us by the browser during
         # HTTPS sessions, if we're using SSL.
         $cookieargs{'-secure'} = 1;
     }
 
-    $cgi->send_cookie(-name => 'Bugzilla_login',
-                      -value => $user->id,
-                      %cookieargs);
-    $cgi->send_cookie(-name => 'Bugzilla_logincookie',
-                      -value => $login_cookie,
-                      %cookieargs);
+    $cgi->send_cookie(
+        -name  => 'Bugzilla_login',
+        -value => $user->id,
+        %cookieargs
+    );
+    $cgi->send_cookie(
+        -name  => 'Bugzilla_logincookie',
+        -value => $login_cookie,
+        %cookieargs
+    );
 }
 
 sub logout {
-    my ($self, $param) = @_;
+    my ( $self, $param ) = @_;
 
-    my $dbh = Bugzilla->dbh;
-    my $cgi = Bugzilla->cgi;
+    my $dbh   = Bugzilla->dbh;
+    my $cgi   = Bugzilla->cgi;
     my $input = Bugzilla->input_params;
     $param = {} unless $param;
     my $user = $param->{user} || Bugzilla->user;
     my $type = $param->{type} || LOGOUT_ALL;
 
-    if ($type == LOGOUT_ALL) {
-        $dbh->do("DELETE FROM logincookies WHERE userid = ?",
-                 undef, $user->id);
+    if ( $type == LOGOUT_ALL ) {
+        $dbh->do( "DELETE FROM logincookies WHERE userid = ?", undef, $user->id );
         return;
     }
 
@@ -107,23 +113,23 @@ sub logout {
     # If a new cookie has been issued during this run, that's the current one.
     # If not, it's the one we've received.
     my @login_cookies;
-    my $cookie = first {$_->name eq 'Bugzilla_logincookie'}
-                       @{$cgi->{'Bugzilla_cookie_list'}};
+    my $cookie = first { $_->name eq 'Bugzilla_logincookie' }
+    @{ $cgi->{'Bugzilla_cookie_list'} };
     if ($cookie) {
-        push(@login_cookies, $cookie->value);
+        push( @login_cookies, $cookie->value );
     }
-    elsif ($cookie = $cgi->cookie('Bugzilla_logincookie')) {
-        push(@login_cookies, $cookie);
+    elsif ( $cookie = $cgi->cookie('Bugzilla_logincookie') ) {
+        push( @login_cookies, $cookie );
     }
 
     # If we are a webservice using a token instead of cookie
     # then add that as well to the login cookies to delete
-    if (my $login_token = $user->authorizer->login_token) {
-        push(@login_cookies, $login_token->{'login_token'});
+    if ( my $login_token = $user->authorizer->login_token ) {
+        push( @login_cookies, $login_token->{'login_token'} );
     }
 
     # Make sure that @login_cookies is not empty to not break SQL statements.
-    push(@login_cookies, '') unless @login_cookies;
+    push( @login_cookies, '' ) unless @login_cookies;
 
     # These queries use both the cookie ID and the user ID as keys. Even
     # though we know the userid must match, we still check it in the SQL
@@ -134,21 +140,19 @@ sub logout {
     # chances? - bbaetz
     map { trick_taint($_) } @login_cookies;
     @login_cookies = map { $dbh->quote($_) } @login_cookies;
-    if ($type == LOGOUT_KEEP_CURRENT) {
-        $dbh->do("DELETE FROM logincookies WHERE " .
-                 $dbh->sql_in('cookie', \@login_cookies, 1) .
-                 " AND userid = ?",
-                 undef, $user->id);
-    } elsif ($type == LOGOUT_CURRENT) {
-        $dbh->do("DELETE FROM logincookies WHERE " .
-                 $dbh->sql_in('cookie', \@login_cookies) .
-                 " AND userid = ?",
-                 undef, $user->id);
-    } else {
+    if ( $type == LOGOUT_KEEP_CURRENT ) {
+        $dbh->do( "DELETE FROM logincookies WHERE " . $dbh->sql_in( 'cookie', \@login_cookies, 1 ) . " AND userid = ?",
+            undef, $user->id );
+    }
+    elsif ( $type == LOGOUT_CURRENT ) {
+        $dbh->do( "DELETE FROM logincookies WHERE " . $dbh->sql_in( 'cookie', \@login_cookies ) . " AND userid = ?",
+            undef, $user->id );
+    }
+    else {
         die("Invalid type $type supplied to logout()");
     }
 
-    if ($type != LOGOUT_KEEP_CURRENT) {
+    if ( $type != LOGOUT_KEEP_CURRENT ) {
         clear_browser_cookies();
     }
